[nmpc][py] move unused code to the archive folder.
