+++ aerial_robot_control/scripts/nmpc/tilt_qd/archive/phys_param_beetle_art.py
