+++ aerial_robot_control/scripts/nmpc/tilt_qd/archive/rh_base.py
